refactor: authv2 -> authutil and remove old auth service
