+++ core/internal/app/access/service/access_service.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	accessmodel "core/internal/app/access/model"
 	accessrepo "core/internal/app/access/repository"
-	"core/internal/pkg/authv2"
+	"core/internal/pkg/authutil"
 	cons "core/internal/pkg/constants"
 	"core/internal/pkg/jwt"
 	rsc "core/internal/pkg/resource"
@@ -77,7 +77,7 @@ func (s *Service) List(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -140,7 +140,7 @@ func (s *Service) Create(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -207,7 +207,7 @@ func (s *Service) Get(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -256,7 +256,7 @@ func (s *Service) Update(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -325,7 +325,7 @@ func (s *Service) Delete(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return &e
