+++ core/internal/app/project/service/project_service.go
@@ -6,7 +6,7 @@ import (
 	projectmodel "core/internal/app/project/model"
 	projectrepo "core/internal/app/project/repository"
 	sdkkeyrepo "core/internal/app/sdkkey/repository"
-	"core/internal/pkg/authv2"
+	"core/internal/pkg/authutil"
 	cons "core/internal/pkg/constants"
 	rsc "core/internal/pkg/resource"
 	"core/internal/pkg/srvenv"
@@ -42,7 +42,7 @@ func (s *Service) List(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -92,7 +92,7 @@ func (s *Service) Create(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -140,7 +140,7 @@ func (s *Service) Get(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -180,7 +180,7 @@ func (s *Service) Update(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -233,7 +233,7 @@ func (s *Service) Delete(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return &e
