+++ core/internal/app/trait/service/trait_service.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	traitmodel "core/internal/app/trait/model"
 	traitrepo "core/internal/app/trait/repository"
-	"core/internal/pkg/authv2"
+	"core/internal/pkg/authutil"
 	cons "core/internal/pkg/constants"
 	rsc "core/internal/pkg/resource"
 	"core/internal/pkg/srvenv"
@@ -35,7 +35,7 @@ func (s *Service) List(
 	defer cancel()
 
 	// Verify access is authorized
-	_, err := authv2.Authorize(s.Senv, atk)
+	_, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -61,7 +61,7 @@ func (s *Service) Create(
 	defer cancel()
 
 	// Verify access is authorized
-	_, err := authv2.Authorize(s.Senv, atk)
+	_, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
