+++ core/internal/app/workspace/service/workspace_service.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	workspacemodel "core/internal/app/workspace/model"
 	workspacerepo "core/internal/app/workspace/repository"
-	"core/internal/pkg/authv2"
+	"core/internal/pkg/authutil"
 	cons "core/internal/pkg/constants"
 	rsc "core/internal/pkg/resource"
 	"core/internal/pkg/srvenv"
@@ -36,7 +36,7 @@ func (s *Service) List(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -82,7 +82,7 @@ func (s *Service) Create(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -114,7 +114,7 @@ func (s *Service) Get(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -150,7 +150,7 @@ func (s *Service) Update(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return nil, &e
@@ -199,7 +199,7 @@ func (s *Service) Delete(
 	defer cancel()
 
 	// Verify access is authorized
-	acc, err := authv2.Authorize(s.Senv, atk)
+	acc, err := authutil.Authorize(s.Senv, atk)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		return &e
